Fixes #146 make verifyCredentials detect suspension.
* Moves verifyCredentials to its own module.
* Modifies setup.js, actions.js, stream.js, and update-users.js to
  use the module version of verifyCredentials instead of an instance
  method on the user object.
* Adds an additional /users/show call in verifyCredentials() to check
  the response for a suspended user flag. If this flag is present the
  account will be set as suspended.

Given the changes in #165 that are staged to be merged to master I did
not change verifyCredentials to return a promise. Happy to submit
another pull req to address that requirement if needed.
