+++ stream.js
@@ -9,7 +9,8 @@ var twitterAPI = require('node-twitter-api'),
     actions = require('./actions'),
     updateUsers = require('./update-users'),
     util = require('./util'),
-    setup = require('./setup');
+    setup = require('./setup'),
+    verifyCredentials = require('./verify-credentials').verifyCredentials;
 
 var twitter = setup.twitter,
     logger = setup.logger,
@@ -201,7 +202,7 @@ function endCallback(user, httpIncomingMessage) {
   var statusCode = httpIncomingMessage.statusCode;
   logger.warn('Ending stream for', user, statusCode);
   if (statusCode === 401 || statusCode === 403) {
-    user.verifyCredentials();
+    verifyCredentials(user);
     delete streams[user.uid];
   } else if (statusCode === 420) {
     // The streaming API will return 420 Enhance Your Calm
@@ -247,7 +248,7 @@ function dataCallback(recipientBtUser, err, data, ret, res) {
     if (data.disconnect.code === 6 ||
         data.disconnect.code === 13 ||
         data.disconnect.code === 14) {
-      recipientBtUser.verifyCredentials();
+      verifyCredentials(recipientBtUser);
     }
   } else if (data.warning) {
     if (data.warning.code === 'FOLLOWS_OVER_LIMIT') {
